Created code to draw SVG circles with random size and color
Pretty excited about this, as I was able to independently
create code that randomized the colors of SVG circles drawn on the screen.

Right now, though, it shows the circle, the fill, and the stroke all
as the SAME RGBA value. I'd like to replace it with a function call
so that it can loop through each time and get different values
for each of the three properties. Saving now just because
I can create a save point where it works.