Using a function to randomize RGBA value of circle fill attribute
So it worked! The current state of the code:

* makes 10 circles of randomized size.
* uses a randomly generated RGBA value for the stroke attribute of
all the circles
* uses a randomly generated RGBA value for the fill attribute of each circle
(so each circle's center is a different color, while all the circles
have the same border color).

A lot of the challenge was concatenating the r, g, b, and a variables
along with text strings to produce the right thing to put as a value
in the .attr(), which needed to look like this: `attr("stroke", "rgba(255,128,12, 0.9)")`
in the DOM.

Leaving a lot of console.log statements in my commit because
these are learning exercises for me and not production code.

🎉 good job! 🎉