All that and it was just a spelling error!
Started Chapter 09 -- the exciting part where we get to animations and
transitions -- and I *could not get* the simple example to work.

I kept getting the error `d3.scalebandwith is not a function`.

That's because I was spelling `bandwidth` wrong. *facepalm*

I looked at the example code and the code I typed again and again and
didn't find the error until I did a diff.